Creating dedicated attachment interaction listener
