+++ vector/src/main/java/im/vector/app/features/media/AttachmentOverlayView.kt
@@ -30,35 +30,38 @@ class AttachmentOverlayView @JvmOverloads constructor(
         context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0
 ) : ConstraintLayout(context, attrs, defStyleAttr), AttachmentEventListener {
 
-    var onShareCallback: (() -> Unit)? = null
-    var onBack: (() -> Unit)? = null
-    var onPlayPause: ((play: Boolean) -> Unit)? = null
-    var videoSeekTo: ((progress: Int) -> Unit)? = null
+    /* ==========================================================================================
+     * Fields
+     * ========================================================================================== */
 
+    var interactionListener: AttachmentInteractionListener? = null
     val views: MergeImageAttachmentOverlayBinding
 
-    var isPlaying = false
+    private var isPlaying = false
+    private var suspendSeekBarUpdate = false
 
-    var suspendSeekBarUpdate = false
+    /* ==========================================================================================
+     * Init
+     * ========================================================================================== */
 
     init {
         inflate(context, R.layout.merge_image_attachment_overlay, this)
         views = MergeImageAttachmentOverlayBinding.bind(this)
         setBackgroundColor(Color.TRANSPARENT)
         views.overlayBackButton.setOnClickListener {
-            onBack?.invoke()
+            interactionListener?.onDismissTapped()
         }
         views.overlayShareButton.setOnClickListener {
-            onShareCallback?.invoke()
+            interactionListener?.onShareTapped()
         }
         views.overlayPlayPauseButton.setOnClickListener {
-            onPlayPause?.invoke(!isPlaying)
+            interactionListener?.onPlayPause(!isPlaying)
         }
 
         views.overlaySeekBar.setOnSeekBarChangeListener(object : SeekBar.OnSeekBarChangeListener {
             override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
                 if (fromUser) {
-                    videoSeekTo?.invoke(progress)
+                    interactionListener?.videoSeekTo(progress)
                 }
             }
 
@@ -72,10 +75,18 @@ class AttachmentOverlayView @JvmOverloads constructor(
         })
     }
 
+    /* ==========================================================================================
+     * Public API
+     * ========================================================================================== */
+
     fun updateWith(counter: String, senderInfo: String) {
         views.overlayCounterText.text = counter
         views.overlayInfoText.text = senderInfo
     }
+
+    /* ==========================================================================================
+     * Specialization
+     * ========================================================================================== */
 
     override fun onEvent(event: AttachmentEvents) {
         when (event) {
